requirements.txt
langchain
langchain-openai
openai
pydantic
fastapi
uvicorn
requests
python-dotenv
streamlit
regex
phonenumbers
email-validator